# Doc/Specifica_OpenAPI.yaml
openapi: 3.0.4
info:
  title: Soccorso Web
  description: |-
    Il sito SoccorsoWeb rappresenta un generico portale per la ricezione e la gestione di richieste di soccorso. Il sito avrà una parte riservata ad amministratori (che configureranno il sistema, smisteranno le richieste e le monitoreranno) e operatori (a cui verranno inviate le richieste e che le gestiranno in prima persona). La parte pubblica del sito, invece, permetterà unicamente di inviare una richieste di soccorso senza la necessità di alcuna registrazione.

    alcuni link utili:
    - [Specifica SWA](https://github.com/SviluppoWebAvanzato-Univaq/SWA_Project_Specifications/blob/main/2024/Progetto_SWA_2024_1_SoccorsoWeb_Services.md)
    - [Specifica WE](https://github.com/WebEngineering-Univaq/WE_Project_Specifications/blob/main/2024/Progetto_WE_2024_1_SoccorsoWeb.md)
  version: 1.0.12
servers:
  - url: http://localhost:8080/SoccorsoWebREST
  - url: http://{regione}.SoccorsoWebREST
    variables:
      regione:
        default: Abruzzo
...
paths:
...
  /rest/richieste/{requid}:
    parameters: 
      - $ref: '#/components/parameters/RichiestaUID'
    delete:
      tags:
        - Richieste
      summary: Annulla richiesta
      description: Eliminazione di una richiesta di soccorso
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/SuccessoEliminazione'
        '401':
          $ref: '#/components/responses/ErroreNonAutorizzato'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ErroreGenerale'
    get:
      tags:
        - Dettagli
      summary: dettagli di una richiesta
      operationId: GetRequest
      security:
        - bearerAuth: []
      responses:
        '200':
          description: dettagli richiesta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Richiesta'
        '401':
          $ref: '#/components/responses/ErroreNonAutorizzato'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ErroreGenerale'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  responses:
...
    SuccessoEliminazione:
      description: Elemento eliminato con successo
    ErroreGenerale:
      description: Errore generale
    ErroreNonAutorizzato:
      description: Token mancante o invalido
    NotFound:
     description: Elemento non trovato
...
  parameters:
...
    RichiestaUID:
      name: requid
      in: path
      required: true
      schema: 
        $ref: '#/components/schemas/UID'
...
  schemas:
    UID:
     type: string
     pattern: ID[a-z0-9-]+
...
    Posizione:
      type: object
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        long:
          type: number
          format: double
          minimum: -180
          maximum: 180
    LvlSuccesso:
      type: number
      format: integer
      default: 0
      minimum: 0
      maximum: 5
    Stato:
      type: string
      enum: [attesa, attiva, in corso, chiusa]
      default: attesa
...
    Richiesta:
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/UID'
        nome:
          type: string
        email:
          type: string
        tempo:
          type: string
          format: date-time
        descrizione:
          type: string
        posizione:
          $ref: '#/components/schemas/Posizione'
        foto:
          type: string
          format: binary
          default: none
        stato:
          $ref: '#/components/schemas/Stato'
        uid_amministratore:
          type: string
          default: none
        lvl_successo:
          $ref: '#/components/schemas/LvlSuccesso'
      required:
        - uid
        - nome
        - contatto
        - descrizione
        - posizione
      additionalProperties: false
...